requirements.txt
pillow
biopython
colorama
matplotlib
matplotlib-inline
pandas
py3dmol
ipython
rdkit
streamlit
requests
stmol == 0.0.7
scikit-bio
neatbio